+++ src/index.ts
@@ -1,0 +1,7 @@
+import { photoData } from './Functions/PhotoData';
+import { translate } from './Functions/Translate';
+import { upload } from './Functions/Upload';
+
+export { upload };
+export { photoData };
+export { translate };