+++ CHANGELOG.md
@@ -1,19 +1,33 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
-## [0.3] - 10/20/2021
+## [0.4] - 10/25/2021
+
 ### Added
-- Better Error System
+
+-   Parsing to responses
+
+## [0.3] - 10/20/2021
+
+### Added
+
+-   Better Error System
 
 ## [0.2] - 10/18/2021
+
 ### Added
-- Translation File Upload
+
+-   Translation File Upload
 
 ## [0.1] - 10/18/2021
+
 ### Added
-- Upload Function
-- PhotoData Function
+
+-   Upload Function
+-   PhotoData Function
 
 [0.1]: https://github.com/subtosharki/fileglass.js/releases/tag/0.1
 [0.2]: https://github.com/subtosharki/fileglass.js/releases/tag/0.2
 [0.3]: https://github.com/subtosharki/fileglass.js/releases/tag/0.3
+[0.4]: https://github.com/subtosharki/fileglass.js/releases/tag/0.4